Move all components connected to Redux to the `containers` folder.
The only component moved was BudgetGrid.
