+++ app/components/BudgetGridRow/style.scss
@@ -1,0 +1,23 @@
+@import 'theme/variables';
+
+.cellLabel {
+  display: none;
+
+  @media only screen and (max-width: $screen-small) {
+    display: flex;
+    align-items: center;
+    flex: 1;
+  }
+}
+
+.cellContent {
+  flex: 1;
+}
+
+.neg {
+  color: $red;
+}
+
+.pos {
+  color: $green;
+}
