+++ app/containers/Budget/index.js
@@ -3,7 +3,7 @@ import * as React from 'react';
 import transactionReducer from 'modules/transactions';
 import categoryReducer from 'modules/categories';
 import { injectAsyncReducers } from 'store';
-import BudgetGrid from 'components/BudgetGrid';
+import BudgetGrid from 'containers/BudgetGrid';
 import Balance from 'containers/Balance';
 
 // inject reducers that might not have been originally there
