+++ app/containers/BudgetGrid/__tests__/index-test.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import renderer from 'react-test-renderer';
-import { BudgetGrid } from '../';
+import { BudgetGrid } from '../index';
 
 // mock nested component
 jest.mock('containers/EntryFormRow');
