+++ app/containers/BudgetGrid/index.js
@@ -5,7 +5,7 @@ import { getTransactions } from 'selectors/transactions';
 import { getCategories } from 'selectors/categories';
 import EntryFormRow from 'containers/EntryFormRow';
 import type { Transaction } from 'modules/transactions';
-import BudgetGridRow from './BudgetGridRow';
+import BudgetGridRow from 'components/BudgetGridRow';
 import styles from './style.scss';
 
 type BudgetGridProps = {
