+++ app/containers/BudgetGrid/style.scss
@@ -54,14 +54,6 @@
     td {
       vertical-align: top;
       text-overflow: ellipsis;
-
-      &.neg {
-        color: $red;
-      }
-
-      &.pos {
-        color: $green;
-      }
     }
   }
 
@@ -95,18 +87,4 @@
       }
     }
   }
-}
-
-.cellLabel {
-  display: none;
-
-  @media only screen and (max-width: $screen-small) {
-    display: flex;
-    align-items: center;
-    flex: 1;
-  }
-}
-
-.cellContent {
-  flex: 1;
 }
